Add files via upload
Updates: Includes inclination angle (meaning more accurate light curves [with the exception of WASP-18b]) and a simple random-based ('Monte-Carlo'-esque) refraction system that can handle around 4000 rays in 160 seconds that uses images and points (rather than matrices for now). The refraction system has a chance to refract, with a random refraction amount, which highlights the position where it refracts. Not accurate as photons (or rays) still need to hit us (maybe use probability to hit us for different angles?).