Add files via upload
(Attempted doing refraction in a 'Monte-Carlo'-sense, but this was at some point dismissed.)
Created a debris cloud by treating it as another 'planet', with intensity scaled by sin(orbitAngle). Input variables are in diameterScale and temperatureScale, relative to the exoplanet (the temperature for a debris cloud doesn't make current sense, especially for albedo). This created shoulders and a very slight intensity at the center of the transit.
The main intensity-overlay calculation skips zeros in the intensity arrays, so the code is easier to understand and programmable.
We found the TESS dataset and can now compare our model to theirs to calculate error.